Hacksmart_Nexora_Navswap_AI/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.4
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
joblib==1.3.2
google-generativeai==0.3.2
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0